+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/method/to_service_method.rs
@@ -1,11 +1,15 @@
 use cdk_framework::act::node::{
     candid::service::Method, node_parts::mode::Mode, CandidType, Param,
 };
-use swc_ecma_ast::{ClassProp, Expr, TsFnOrConstructorType, TsFnType, TsType};
+use swc_ecma_ast::{ClassProp, Expr, TsFnOrConstructorType, TsFnParam, TsFnType, TsType};
 
 use crate::{
     errors::service_method::ParseError,
-    ts_ast::{azle_type::AzleType, source_map::SourceMapped, GetName},
+    ts_ast::{
+        azle_type::AzleType,
+        source_map::SourceMapped,
+        traits::{GetName, GetTsType},
+    },
 };
 
 impl SourceMapped<'_, ClassProp> {
@@ -134,5 +138,36 @@ impl SourceMapped<'_, ClassProp> {
             },
             None => return Err(ParseError::MissingTypeAnnotation),
         }
+    }
+}
+
+impl SourceMapped<'_, TsFnType> {
+    pub fn build_act_fn_params(&self) -> Vec<Param> {
+        self.params
+            .iter()
+            .map(|param| match param {
+                TsFnParam::Ident(identifier) => {
+                    let name = identifier.get_name().to_string();
+                    let candid_type = match &identifier.type_ann {
+                        Some(ts_type_ann) => {
+                            let azle_type =
+                                AzleType::from_ts_type(ts_type_ann.get_ts_type(), self.source_map);
+                            azle_type.to_data_type()
+                        }
+                        None => panic!("Function parameters must have a type"),
+                    };
+                    Param { name, candid_type }
+                }
+                TsFnParam::Array(_) => {
+                    panic!("Array destructuring in parameters is unsupported at this time")
+                }
+                TsFnParam::Rest(_) => {
+                    panic!("Rest parameters are not supported at this time")
+                }
+                TsFnParam::Object(_) => {
+                    panic!("Object destructuring in parameters is unsupported at this time")
+                }
+            })
+            .collect()
     }
 }
