+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/mod.rs
@@ -1,5 +1,4 @@
 mod ts_fn_or_constructor_type;
-mod ts_fn_type;
 mod ts_type;
 mod ts_type_element;
