--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type_ann.rs
@@ -1,9 +1,0 @@
-use swc_ecma_ast::TsTypeAnn;
-
-use super::traits::GetTsType;
-
-impl GetTsType for TsTypeAnn {
-    fn get_ts_type(&self) -> swc_ecma_ast::TsType {
-        *self.type_ann.clone()
-    }
-}
